Add node-common lib
