+++ core-rust/core-rust/src/lib.rs
@@ -3,6 +3,10 @@
 /// I truly have no idea why this works, but it does.
 #[no_mangle]
 fn export_extern_functions() {
+    // node-common
+    node_common::jni::addressing::export_extern_functions();
+    node_common::jni::scrypto_constants::export_extern_functions();
+
     // state-manager
     state_manager::jni::mempool::export_extern_functions();
     state_manager::jni::state_computer::export_extern_functions();
@@ -10,8 +14,6 @@ fn export_extern_functions() {
     state_manager::jni::transaction_builder::export_extern_functions();
     state_manager::jni::transaction_store::export_extern_functions();
     state_manager::jni::vertex_store_recovery::export_extern_functions();
-    state_manager::jni::addressing::export_extern_functions();
-    state_manager::jni::scrypto_constants::export_extern_functions();
 
     // core-api-server
     core_api_server::jni::export_extern_functions();
