+++ core-rust/node-common/Cargo.toml
@@ -1,12 +1,9 @@
 [package]
-name = "core-api-server"
+name = "node-common"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-node-common = { path = "../node-common" }
-state-manager = { path = "../state-manager" }
-
 sbor = { workspace = true }
 transaction = { workspace = true }
 radix-engine-common = { workspace = true }
@@ -25,11 +22,13 @@ tokio = { workspace = true }
 parking_lot = { workspace = true }
 prometheus = { workspace = true }
 blake2 = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
+bech32 = { version = "0.9.0", default-features = false }
 
-chrono = { version = "0.4.23", default-features = false, features = ["std"] }
-hex = { version = "0.4.3", default-features = false }
-futures = { version = "0.3" }
-axum = { version = "0.6.6", features = ["http1", "json"] }
-hyper = { version = "0.14.20", features = ["server", "http1"] }
+tracing-opentelemetry = { version = "0.18" }
+tracing-subscriber = { version = "0.3.15" }
+opentelemetry = { version = "0.18", default-features = false, features = [
+  "rt-tokio",
+  "trace",
+] }
+opentelemetry-jaeger = { version = "0.17", features = ["rt-tokio"] }
+
