+++ core-rust/node-common/src/environment.rs
