+++ core-rust/node-common/src/utils.rs
