+++ core-rust/state-manager/Cargo.toml
@@ -4,6 +4,8 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+node-common = { path = "../node-common" }
+
 sbor = { workspace = true }
 transaction = { workspace = true }
 radix-engine-common = { workspace = true }
@@ -22,20 +24,11 @@ tokio = { workspace = true }
 parking_lot = { workspace = true }
 prometheus = { workspace = true }
 blake2 = { workspace = true }
-lazy_static = { workspace = true }
 
 rocksdb = { version = "0.19.0" }
 hex = { version = "0.4.3", default-features = false }
-bech32 = { version = "0.9.0", default-features = false }
 lru = { version = "0.8.0", default-features = false }
 slotmap = { version = "1.0.6" }
 im = { version = "15.1.0" }
 rand = { version = "0.8.5" }
-tracing-opentelemetry = { version = "0.18" }
-tracing-subscriber = { version = "0.3.15" }
-opentelemetry = { version = "0.18", default-features = false, features = [
-  "rt-tokio",
-  "trace",
-] }
-opentelemetry-jaeger = { version = "0.17", features = ["rt-tokio"] }
 enum_dispatch = { version = "0.3.11" }
