+++ core-rust/state-manager/src/jni/state_computer.rs
@@ -75,10 +75,10 @@ use radix_engine::types::*;
 use radix_engine_queries::query::ResourceAccounter;
 use std::ops::Deref;
 
-use crate::jni::common_types::JavaHashCode;
 use crate::jni::state_manager::JNIStateManager;
-use crate::jni::utils::*;
 use crate::query::StateManagerSubstateQueries;
+use node_common::java::*;
+use node_common::jni::utils::*;
 
 use crate::types::{CommitRequest, PrepareRequest, PrepareResult};
 use crate::{CommitError, NextEpoch};
